Clean up some of protocol abstractions - Renamed PipeConnection to DuplexPipe - Removed MemoryPool from ConnectionContext
